requirements.txt
streamlit
pandas
numpy
scikit-learn
pillow
pytesseract
joblib
